split ZkVmHost into Prover and Verifier
