+++ adapters/sp1/src/host.rs
@@ -1,13 +1,7 @@
 use std::fmt;
 
-use serde::{de::DeserializeOwned, Serialize};
-use sp1_sdk::{network::FulfillmentStrategy, HashableKey, ProverClient, SP1ProvingKey};
-use zkaleido::{
-    ProofType, PublicValues, VerificationKey, VerificationKeyCommitment, ZkVmError, ZkVmHost,
-    ZkVmInputBuilder, ZkVmResult,
-};
-
-use crate::{input::SP1ProofInputBuilder, proof::SP1ProofReceipt};
+use sp1_sdk::{HashableKey, ProverClient, SP1ProvingKey};
+use zkaleido::ZkVmHost;
 
 /// A host for the `SP1` zkVM that stores the guest program in ELF format.
 /// The `SP1Host` is responsible for program execution and proving
@@ -38,107 +32,7 @@ impl SP1Host {
     }
 }
 
-impl ZkVmHost for SP1Host {
-    type Input<'a> = SP1ProofInputBuilder;
-    type ZkVmProofReceipt = SP1ProofReceipt;
-    fn prove_inner<'a>(
-        &self,
-        prover_input: <Self::Input<'a> as ZkVmInputBuilder<'a>>::Input,
-        proof_type: ProofType,
-    ) -> ZkVmResult<SP1ProofReceipt> {
-        // If the environment variable "ZKVM_MOCK" is set to "1" or "true" (case-insensitive),
-        // then set "SP1_PROVER" to "mock". This effectively enables the mock mode in the SP1
-        // prover.
-        if std::env::var("ZKVM_MOCK")
-            .map(|v| v == "1" || v.to_lowercase() == "true")
-            .unwrap_or(false)
-        {
-            std::env::set_var("SP1_PROVER", "mock");
-        }
-
-        // Handle network proving with custom strategy
-        if std::env::var("SP1_PROOF_STRATEGY").unwrap_or_default() == "private_cluster" {
-            let prover_client = ProverClient::builder().network().build();
-
-            let network_prover_builder = prover_client
-                .prove(&self.proving_key, &prover_input)
-                .strategy(FulfillmentStrategy::Reserved);
-
-            let network_prover = match proof_type {
-                ProofType::Compressed => network_prover_builder.compressed(),
-                ProofType::Core => network_prover_builder.core(),
-                ProofType::Groth16 => network_prover_builder.groth16(),
-            };
-
-            let proof_info = network_prover
-                .run()
-                .map_err(|e| ZkVmError::ProofGenerationError(e.to_string()))?;
-
-            return Ok(proof_info.into());
-        }
-
-        let client = ProverClient::from_env();
-        let mut prover = client.prove(&self.proving_key, &prover_input);
-
-        prover = match proof_type {
-            ProofType::Compressed => prover.compressed(),
-            ProofType::Core => prover.core(),
-            ProofType::Groth16 => prover.groth16(),
-        };
-
-        let proof_info = prover
-            .run()
-            .map_err(|e| ZkVmError::ProofGenerationError(e.to_string()))?;
-
-        Ok(proof_info.into())
-    }
-
-    fn execute<'a>(
-        &self,
-        prover_input: <Self::Input<'a> as ZkVmInputBuilder<'a>>::Input,
-    ) -> ZkVmResult<PublicValues> {
-        let client = ProverClient::from_env();
-
-        let (output, _) = client
-            .execute(self.get_elf(), &prover_input)
-            .run()
-            .map_err(|e| ZkVmError::ExecutionError(e.to_string()))?;
-
-        let public_values = PublicValues::new(output.to_vec());
-
-        Ok(public_values)
-    }
-
-    fn extract_serde_public_output<T: Serialize + DeserializeOwned>(
-        public_values: &PublicValues,
-    ) -> ZkVmResult<T> {
-        let public_params: T = bincode::deserialize(public_values.as_bytes())
-            .map_err(|e| ZkVmError::OutputExtractionError { source: e.into() })?;
-        Ok(public_params)
-    }
-
-    fn get_elf(&self) -> &[u8] {
-        &self.proving_key.elf
-    }
-
-    fn get_verification_key(&self) -> VerificationKey {
-        let verification_key = bincode::serialize(&self.proving_key.vk).unwrap();
-        VerificationKey::new(verification_key)
-    }
-
-    fn get_verification_key_commitment(&self) -> VerificationKeyCommitment {
-        VerificationKeyCommitment::new(self.proving_key.vk.hash_u32())
-    }
-
-    fn verify_inner(&self, proof: &SP1ProofReceipt) -> ZkVmResult<()> {
-        let client = ProverClient::from_env();
-        client
-            .verify(proof.as_ref(), &self.proving_key.vk)
-            .map_err(|e| ZkVmError::ProofVerificationError(e.to_string()))?;
-
-        Ok(())
-    }
-}
+impl ZkVmHost for SP1Host {}
 
 impl fmt::Debug for SP1Host {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
